requirements.txt
future
PyPDF2
opencv-python
pdf2image
numpy
Pillow